Migrated Class<?> getType() to ClassO getType(), this means that reflection Class was replaced with domain ClassO almost everywhere, this is first step to wrapping reflection api

+++ src/main/java/src/parsing/domain/constantValues/IntLiteral.java
@@ -56,7 +56,7 @@ public class IntLiteral extends Value {
 
     @Override
     public ClassO getType() {
-        return ClassO.FLOAT;
+        return ClassO.INT;
     }
 
 }
